// Frontend/src/swagger/model/signUp.ts
/**
 * Expense Tracker API
 * REST API endpoints for Expense Tracker App
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface SignUp { 
    name: string;
    email: string;
    password: string;
    isPremium: boolean;
}